Added optimizeForSpeed to getScreenshot(). That seems to be doubling the throughput of my program.
Added some error checking to getScreenShot().
Any time it fails, send the exception to the console.
If it fails once, try again.
If it fails twice in a row, return undefined.
If the main program receives undefined, it will report how far it got, then it will do a graceful shutdown.

Previously I was not catching any exceptions.
Deno would see an uncaught exception and halt immediately.
The output file would be unusable.

Add the slurpAll option.
That asks the remote web page how many frames it would like to display.
Then it asks for those frames one at a time.

Disabled assertValidT()
Some scripts still expect inputs between 0 and 1.
However, sometimes I want to send a frame number that is an integer, and this program might not know the valid range.

Added the dx demo script.
That produced https://youtu.be/THZZlEpo684