Works with parabola-tangent-line.html
This is part of the "better derivative" script.

Added a new option to slurpAll.
The remote html file can specify how many seconds it wants to run for.
Previously the html could only specify the number of frames.
Now the html program doesn't have to know or care about the frame rate.